[PATCH] Update help text to indicate --mem-usage is supported on archs other than x86_64
Commit 8449bda73ab14516d4bf81d29503c1ea203bb865 ("Documentation:
Update documentation regarding --mem-usage' option"), updated the
makedumpfile man page to indicate that this option is now
supported on x86_64, arm64, ppc64 and s390x.

However, the help text for makedumpfile (which one can see
by running 'makedumpfile --help'), still reflects the old
support status, i.e. this option is only supported on x86_64.

This patch changes the help text to be in sync with the man
page text.

Signed-off-by: Bhupesh Sharma <bhsharma@redhat.com>