[PATCH] x86_64: fix get_kaslr_offset_x86_64() to return kaslr_offset correctly
Currently, the get_kaslr_offset_x86_64() function has the following
condition to return info->kaslr_offset, but it is wrong, and it can
return 0 for kernel text addresses if info->kaslr_offset is small.

    if (vaddr >= __START_KERNEL_map &&
                    vaddr < __START_KERNEL_map + info->kaslr_offset)

Consequently, kernel text symbols in erase config could be resolved
wrongly, and makedumpfile fails to vtop with the following message
or erases unintended data.

    __vtop4_x86_64: Can't get a valid pmd_pte.

To fix this, use NUMBER(KERNEL_IMAGE_SIZE) in vmcoreinfo if any,
otherwise use the hard-coded value (1 GiB) for KASLR, which has not
been changed from the initial KASLR implementation.

Signed-off-by: Kazuhito Hagio <k-hagio@ab.jp.nec.com>