[PATCH] x86_64: Fix incorrect exclusion by -e option with KASLR
The -e option uses info->vmemmap_start for creating a table to determine
the positions of page structures that should be excluded, but it is a
hardcoded value even with KASLR-enabled vmcore.  As a result, the option
excludes incorrect pages from it.

To fix this, get the vmemmap start address from info->mem_map_data.

Signed-off-by: Kazuhito Hagio <k-hagio@ab.jp.nec.com>