[PATCH] Fix divide by zero in print_report()
If info->max_mapnr and pfn_memhole are equal, we divide by zero when
trying determine the 'shrinking' value.

On the system I saw this error, we arrived at this function with
info->max_mapnr:0x0000000001080000 pfn_memhole:0x0000000001080000

Change the code to only print out the shrinking value if it makes sense.

Signed-off-by: Dave Jones <davej@codemonkey.org.uk>