[PATCH] Fix off-by-one issue in exclude_nodata_pages()
When building a dump bitmap (2nd bitmap) for the ELF dump, the last pfn
of the cycle is always ignored in exclude_nodata_pages() function due to
off-by-one error on cycle boundary check. Thus, the respective bit of
the bitmap is never cleared.

That can lead to the error when such a pfn should not be dumpable (e.g.
the last pfn of the ELF-load of zero filesize). Based on the bit in the
bitmap the page is treated as dumpable in write_elf_pages_cyclic() function
and the follow on error is triggered in write_elf_load_segment() function
due to the failing sanity check of paddr_to_offset2():

   $ makedumpfile -E dump.elf dump.elf.E
   Checking for memory holes                         : [100.0 %] |
   write_elf_load_segment: Can't convert physaddr(7ffff000) to an offset.
   makedumpfile Failed.

Signed-off-by: Mikhail Zaslonko <zaslonko@linux.ibm.com>