[PATCH] Pass 0 to get_kaslr_offset() in find_kaslr_offsets()
Currently SYMBOL(_stext) is passed to get_kaslr_offset() in
find_kaslr_offsets(), but it is always zero, because it has not
been set yet at the time.

On the other hand, the vaddr argument of get_kaslr_offset() is
only used to decide whether to return a KASLR offset or not, but
the return value is not used in find_kaslr_offsets().

Therefore, passing SYMBOL(_stext) is meaningless and confusing,
so let's pass it 0 explicitly to avoid confusion.

Reported-by: Lianbo Jiang <lijiang@redhat.com>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>