[PATCH] Show write byte size in report messages
Show write byte size in report messages.  This value can be different
from the size of the actual file because of some holes on dumpfile
data structure.

  $ makedumpfile --show-stats -l -d 1 vmcore dump.ld1
  ...
  Total pages     : 0x0000000000080000
  Write bytes     : 377686445
  ...
  # ls -l dump.ld1
  -rw------- 1 root root 377691573 Feb  4 16:28 dump.ld1

Note that this value should not be used with /proc/kcore to determine
how much disk space is needed for crash dump, because the real memory
usage when a crash occurs can vary widely.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>