[PATCH v2 3/3] Add -L option to limit output file size
This option can be used to ensure that a certain amount of free space is
preserved.  It is useful when the output of makedumpfile is on the root
filesystem and some services fail to start at boot if there is no space
left.

Note that in some cases the limit can be triggered before the output file
reaches the specified size because makedumpfile seeks and tries to write
beyond the end of the file.

Signed-off-by: Benjamin Poirier <bpoirier@nvidia.com>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>