[PATCH 1/2] Fix bad file descriptor error when using --dry-run
When running makedumpfile with the --dry-run option it fails with

    write_and_check_space: Can't seek the dump file(vmcore). Bad file descriptor

This is because for --dry-run no dump file is created and a dummy file
descriptor of -1 is used. Thus the lseek in write_and_check_space will
fail. Fix this by treating a dry run as if writing to STDOUT.

Fixes: f0cfa86 ("[PATCH v2 3/3] Add -L option to limit output file size")
Signed-off-by: Philipp Rudo <prudo@redhat.com>