[PATCH 2/2] Fix --dry-run for incomplete dumps
When running out of space during a dry run, e.g. by limiting the output
size using the -L option, makedumpfile fails with

    Can't write the dump file(vmcore). Size limit(104857600) reached.
    get_nr_pages: Can't seek end of the dump file(vmcore).
    __read_disk_dump_header: Can't open a file(vmcore). No such file or directory

This is because for --dry-run no dump file is created and a dummy file
descriptor of -1 is used. Thus the file operations in get_nr_pages and
check_and_modify_*_headers fail. Fix the first error by using
write_bytes as file size for the output file and the second one by
always returning TRUE for check_and_modify_headers.

Fixes: 3422e1d ("[PATCH 1/2] Add --dry-run option to prevent writing the dumpfile")
Signed-off-by: Philipp Rudo <prudo@redhat.com>