[PATCH 03/14] fix buffer overflow in DumpInfo in read_vmcoreinfo_basic_info()
While the kernel limits the osrelease to 65 bytes makedumpfile cannot
assume that is always the case. For example a corrupted dump could
have the terminating \0 byte of the osrelease field in the vmcoreinfo
overwritten. In that case strcpy() causes a buffer overflow in the
DumpInfo. Fix this by using memcpy() instead. This might truncate
the osrelease. (snprintf() could not be used due to a gdb-11
compilation warning.)

Reproducer:
  $ makedumpfile -d 31 vmcore.elf.corrupted foo

Fixes: 356896a ("Initial revision")
Signed-off-by: Philipp Rudo <prudo@redhat.com>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>