[PATCH 13/15] remove variable length array in write_kdump_pages_cyclic()
Variable length arrays can be problematic. For example with C11 they got
downgraded to an optional feature and thus might not be supported in all
environments. Furthermore they don't allow error checks whether the
allocation succeeded. If it fails this causes undefined behavior. This is
especially problematic when its size is read from a dump, like it is done
for info->page_size, as it might have been corrupted. Thus get rid of
variable length arrays and allocate them on the heap instead.

Signed-off-by: Philipp Rudo <prudo@redhat.com>