[PATCH] Makefile: Avoid installing files in /etc
Some Linux distros rely on having files in /etc (Debian, et al.),
others in *not* having them in such directory (like Arch Linux).
Fact is that the former distros have no issues in installing files
elsewhere, whereas Arch has issues in installing files on /etc,
specially if such files are present there when user try to reinstall
the packages - it fails. Arch packaging doesn't try to be smart as dpkg
for example (that compares config files to determine if user changed that).

With all of that said, this patch moves the sample conf file to
/usr/share, a move that is well-tolerated in all distros and shouldn't
cause regressions in packaging. Also, if some Linux distribution likes
the idea of adding files in /etc, they can tune it in their packaging
configuration scripts for makedumpfile, but we shouldn't have that
as default in the Makefile.

[ kh: changed makedumpfile.spec accordingly. ]

Cc: Coiby Xu <coxu@redhat.com>
Cc: Kazuhito Hagio <k-hagio-ab@nec.com>
Cc: Leonidas Spyropoulos <artafinde@archlinux.org>
Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>