[PATCH] sadump: fix failure of reading memory when 5-level paging is enabled
makedumpfile fails as follows for memory dumps collected by sadump
when 5-level paging is enabled on the corresponding systems:

    # makedumpfile -l -d 31 -x ./vmlinux ./dump.sadump dump.sadump-ld31
    __vtop4_x86_64: Can't get a valid pgd.
    ...snip...
    __vtop4_x86_64: Can't get a valid pgd.
    calc_kaslr_offset: failed to calculate kaslr_offset and phys_base; default to 0
    __vtop4_x86_64: Can't get a valid pgd.
    readmem: Can't convert a virtual address(ffffffff82fce960) to physical address.
    readmem: type_addr: 0, addr:ffffffff82fce960, size:1024
    cpu_online_mask_init: Can't read cpu_online_mask memory.

    makedumpfile Failed.

This is because 5-level paging support has not been done yet for
sadump; the work of the 5-level paging support was done by the commit
30a3214a7193e94c551c0cebda5918a72a35c589 (PATCH 4/4 arch/x86_64: Add
5-level paging support) but that was focused on the core part only.

Having said that, most of things has already been finished in the
commit. What needs to be newly added for sadump is just how to check
if 5-level paging is enabled for a given memory dump.

For that purpose, let's refer to CR4.LA57, bit 12 of CR4, representing
whether 5-level paging is enabled or not. We can do this because
memory dumps collected by sadump have SMRAM as note information and
they include CR4 together with the other control registers.

Signed-off-by: HATAYAMA Daisuke <d.hatayama@fujitsu.com>