[PATCH] ppc64: read cur_mmu_type from vmcoreinfo
Currently makedumpfile depends on reading the 'cur_cpu_spec' kernel
symbol to get the current MMU type on PowerPC64.

The disadvantage with this approach was that it depends on bit '0x40'
('MMU_FTR_TYPE_RADIX') being set in 'cur_cpu_spec->mmu_features',
which implies kernel developers have to be careful of modifying
MMU_FTR_* defines.

Instead a more stable approach was suggested by contributors in [1], to
pass information about the MMU type in vmcoreinfo itself, instead of
depending on the MMU_FTR_* defines.

This was implemented in linux kernel in:
  commit 36e826b568e4 ("powerpc/vmcore: Add MMU information to vmcoreinfo")

With this commit, if RADIX_MMU is there in the vmcoreinfo, we prefer it
to get current mmu type, instead of 'cur_cpu_spec'.
On older kernels, where RADIX_MMU number is not there, makedumpfile will
simply fall back to using 'cur_cpu_spec'.

The earlier defines for 'RADIX_MMU' have been renamed to 'MMU_TYPE_RADIX'
which avoids conflict with the vmcoreinfo string 'RADIX_MMU', as well as
being more clear about the value 0x40 with a comment about MMU_FTR_TYPE_RADIX.

[1] https://lore.kernel.org/linuxppc-dev/87v8c3m70t.fsf@mail.lhotse/

Signed-off-by: Aditya Gupta <adityag@linux.ibm.com>