[PATCH] Improve message readability and fix typos
This patch enhances the readability of several messages printed by
adding a space after a comma that is not followed by a newline. These
changes ensure that messages are easier to read and understand.

For example:
Before: "The kernel doesn't support mmap(),read() will be used instead."
After: "The kernel doesn't support mmap(), read() will be used instead."

Additionally, this patch fixes minor typos in the messages, further
improving clarity and correctness.

Signed-off-by: Seiji Nishikawa <snishika@redhat.com>